Hide arcana behind pimpl in AppRuntime
Move arcana types (cancellation_source, manual_dispatcher), thread,
env, suspension lock, and the Append template into a private Impl
class defined in AppRuntime.cpp. This keeps arcana out of the public
header so consumers don't need arcana's include directories.

The worker thread is joined before Impl is destroyed, preserving the
same lifetime guarantees as before.
