+++ Core/AppRuntime/Include/Babylon/AppRuntime.h
@@ -6,16 +6,9 @@
 
 #include <napi/utilities.h>
 
-#include <arcana/threading/cancellation.h>
-#include <arcana/threading/dispatcher.h>
-
 #include <memory>
 #include <functional>
 #include <exception>
-#include <optional>
-#include <mutex>
-#include <thread>
-#include <type_traits>
 
 namespace Babylon
 {
@@ -48,23 +41,6 @@ namespace Babylon
         static void BABYLON_API DefaultUnhandledExceptionHandler(const Napi::Error& error);
 
     private:
-        template<typename CallableT>
-        void Append(CallableT callable)
-        {
-            if constexpr (std::is_copy_constructible<CallableT>::value)
-            {
-                m_dispatcher.queue([this, callable = std::move(callable)]() {
-                    callable(m_env.value());
-                });
-            }
-            else
-            {
-                m_dispatcher.queue([this, callablePtr = std::make_shared<CallableT>(std::move(callable))]() {
-                    (*callablePtr)(m_env.value());
-                });
-            }
-        }
-
         // These three methods are the mechanism by which platform- and JavaScript-specific
         // code can be "injected" into the execution of the JavaScript thread. These three
         // functions are implemented in separate files, thus allowing implementations to be
@@ -84,10 +60,8 @@ namespace Babylon
         void Execute(Dispatchable<void()> callback);
 
         Options m_options;
-        std::optional<Napi::Env> m_env{};
-        std::optional<std::scoped_lock<std::mutex>> m_suspensionLock{};
-        arcana::cancellation_source m_cancelSource{};
-        arcana::manual_dispatcher<128> m_dispatcher{};
-        std::thread m_thread;
+
+        class Impl;
+        std::unique_ptr<Impl> m_impl;
     };
 }
